Change HTTPS tests to use "localhost"
The tests were using "android.com" as the client socket hostname, but
the server was presenting a certificate for "localhost"  This led to a
problem when enabling endpoint algorithm checking.

This could be made more clever with serialization tricks on the
InetAddress, but that would be less portable due to SecurityManager.

Test: cts-tradefed run cts -m CtsLibcoreOkHttpTestCases -a arm64-v8a
Change-Id: I8ed93bb289dc2d26ae885ce97e67766f0db48fae
Signed-off-by: mydongistiny <jaysonedson@gmail.com>